Move Product Tag Client into Tags directory
